requirements.txt
openai==1.3.7
python-dotenv==1.0.0
pandas==2.0.1
requests==2.31.0
